Use restrucured GetCaller and PostCaller.
